gtk3: Workaround an issue with the dark theme in synapic
The "Quick filter" entry in the synaptic toolbar uses a hard coded yellow color to indicate when it's searching. Because of how it's done the only way to make the text visible in the dark theme is to use a light colored entry. Use as specific a selector as possible and hope it doesn't show up somewhere else.